Refactor event log utilities
Closes #5
