+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pix-framework"
-version = "0.10.8"
+version = "0.10.9"
 description = "Process Improvement Explorer Framework contains process discovery and improvement modules of the Process Improvement Explorer project."
 authors = [
     "David Chapela de la Campa <david.chapela.delacampa@gmail.com>",
