+++ src/pix_framework/discovery/case_arrival.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from pix_framework.calendar.resource_calendar import RCalendar
 from pix_framework.discovery.calendar_factory import CalendarFactory
-from pix_framework.log_ids import EventLogIDs
+from pix_framework.io.event_log import EventLogIDs
 from pix_framework.statistics.distribution import (
     get_best_fitting_distribution,
     get_observations_histogram,
