+++ tests/pix_framework/discovery/profile_resource_model_profile_differentiated.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 from pix_framework.discovery.resource_calendars import CalendarDiscoveryParams, CalendarType
 from pix_framework.discovery.resource_model import ResourceModel, discover_resource_model
-from pix_framework.input import read_csv_log
-from pix_framework.log_ids import PROSIMOS_LOG_IDS
+from pix_framework.io.event_log import read_csv_log
+from pix_framework.io.event_log import PROSIMOS_LOG_IDS
 
 assets_dir = Path(__file__).parent.parent / "assets"
