+++ tests/pix_framework/discovery/test_resource_activity_performances.py
@@ -7,8 +7,8 @@ from pix_framework.discovery.resource_calendars import discover_resource_calenda
     CalendarType
 from pix_framework.discovery.resource_profiles import discover_undifferentiated_resource_profile, \
     discover_differentiated_resource_profiles, discover_pool_resource_profiles
-from pix_framework.input import read_csv_log
-from pix_framework.log_ids import APROMORE_LOG_IDS
+from pix_framework.io.event_log import read_csv_log
+from pix_framework.io.event_log import APROMORE_LOG_IDS
 
 assets_dir = Path(__file__).parent.parent / "assets"
