+++ tests/pix_framework/discovery/test_resource_model.py
@@ -6,8 +6,8 @@ from pix_framework.discovery.resource_activity_performances import ActivityResou
 from pix_framework.discovery.resource_calendars import CalendarDiscoveryParams, CalendarType
 from pix_framework.discovery.resource_model import ResourceModel, discover_resource_model
 from pix_framework.discovery.resource_profiles import ResourceProfile
-from pix_framework.input import read_csv_log
-from pix_framework.log_ids import APROMORE_LOG_IDS, PROSIMOS_LOG_IDS
+from pix_framework.io.event_log import APROMORE_LOG_IDS, read_csv_log
+from pix_framework.io.event_log import PROSIMOS_LOG_IDS
 
 assets_dir = Path(__file__).parent.parent / "assets"
