+++ tests/pix_framework/discovery/test_resource_model_benchmark.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import pytest
 from pix_framework.discovery.resource_calendars import CalendarDiscoveryParams, CalendarType
 from pix_framework.discovery.resource_model import ResourceModel, discover_resource_model
-from pix_framework.input import read_csv_log
-from pix_framework.log_ids import PROSIMOS_LOG_IDS
+from pix_framework.io.event_log import read_csv_log
+from pix_framework.io.event_log import PROSIMOS_LOG_IDS
 
 assets_dir = Path(__file__).parent.parent / "assets"
