+++ tests/pix_framework/discovery/test_resource_pools.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import pytest
 
 from pix_framework.discovery.resource_pools import discover_resource_pools
-from pix_framework.input import read_csv_log
-from pix_framework.log_ids import APROMORE_LOG_IDS
+from pix_framework.io.event_log import read_csv_log
+from pix_framework.io.event_log import APROMORE_LOG_IDS
 
 assets_dir = Path(__file__).parent.parent / "assets"
